+++ database/migrations/2025_06_25_093425_create_air_conditioners_table.php
@@ -93,7 +93,7 @@ return new class extends Migration
             // Refrigerant system
             $table->string('refrigerant_kind')->nullable();
             $table->string('needs_to_be_completed')->nullable();
-            $table->integer('refrigerant_amount')->nullable();
+            $table->decimal('refrigerant_amount', 8, 2)->nullable();
 
             // Power consumption and efficiency
             $table->integer('rated_voltage')->nullable();
